# Optimize coargsort for numeric arrays